Refactor server plugins to conform to latest API changes
s
