+++ server/glsp-ecore/src/main/java/com/eclipsesource/glsp/ecore/EcoreOperationHandlerProvider.java
@@ -10,20 +10,21 @@
  ******************************************************************************/
 package com.eclipsesource.glsp.ecore;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
 
 import com.eclipsesource.glsp.api.operations.OperationHandler;
-import com.eclipsesource.glsp.api.operations.OperationHandlerProvider;
-import com.eclipsesource.glsp.server.handlers.DeleteHandler;
-import com.eclipsesource.glsp.server.handlers.MoveNodeHandler;
+import com.eclipsesource.glsp.api.provider.OperationHandlerProvider;
+import com.eclipsesource.glsp.server.operationhandler.DeleteHandler;
+import com.eclipsesource.glsp.server.operationhandler.MoveNodeHandler;
+
 
 public class EcoreOperationHandlerProvider implements OperationHandlerProvider {
 
-	List<OperationHandler> operationHandlers = new ArrayList<>();
+	Set<OperationHandler> operationHandlers = new HashSet<>();
 
 	@Override
-	public List<OperationHandler> getOperationHandlers() {
+	public Set<OperationHandler> getOperationHandlers() {
 //		operationHandlers.add(new CreateAutomatedTaskHandler());
 //		operationHandlers.add(new CreateManualTaskHandler());
 //		operationHandlers.add(new CreateDecisionNodeHandler());
