+++ server/glsp-ecore/src/main/java/com/eclipsesource/glsp/ecore/EcoreServerRuntimeModule.java
@@ -12,21 +12,26 @@ package com.eclipsesource.glsp.ecore;
 
 import com.eclipsesource.glsp.api.factory.ModelFactory;
 import com.eclipsesource.glsp.api.factory.PopupModelFactory;
-import com.eclipsesource.glsp.api.listener.GraphicalModelExpansionListener;
-import com.eclipsesource.glsp.api.listener.GraphicalModelSelectionListener;
-import com.eclipsesource.glsp.api.listener.ModelElementOpenListener;
+import com.eclipsesource.glsp.api.model.ModelElementOpenListener;
+import com.eclipsesource.glsp.api.model.ModelExpansionListener;
+import com.eclipsesource.glsp.api.model.ModelSelectionListener;
+import com.eclipsesource.glsp.api.model.ModelTypeConfiguration;
 import com.eclipsesource.glsp.api.operations.OperationConfiguration;
-import com.eclipsesource.glsp.api.operations.OperationHandlerProvider;
+import com.eclipsesource.glsp.api.provider.OperationHandlerProvider;
 import com.eclipsesource.glsp.ecore.diagram.EcoreModelFactory;
-import com.eclipsesource.glsp.server.di.ServerModule;
-
+import com.eclipsesource.glsp.server.ServerModule;
 
 public class EcoreServerRuntimeModule extends ServerModule {
-	
+
 	@Override
 	public Class<? extends ModelFactory> bindModelFactory() {
 //		return EcoreSprottyFileModelFactory.class;
 		return EcoreModelFactory.class;
+	}
+
+	@Override
+	protected Class<? extends ModelTypeConfiguration> bindModelTypeConfiguration() {
+		return EcoreModelTypeConfiguration.class;
 	}
 
 	@Override
@@ -35,7 +40,7 @@ public class EcoreServerRuntimeModule extends ServerModule {
 	}
 
 	@Override
-	public Class<? extends GraphicalModelSelectionListener> bindGraphicalModelSelectionListener() {
+	public Class<? extends ModelSelectionListener> bindModelSelectionListener() {
 		return EcoreServerListener.class;
 	}
 
@@ -45,7 +50,7 @@ public class EcoreServerRuntimeModule extends ServerModule {
 	}
 
 	@Override
-	public Class<? extends GraphicalModelExpansionListener> bindGraphicalModelExpansionListener() {
+	public Class<? extends ModelExpansionListener> bindModelExpansionListener() {
 		return EcoreServerListener.class;
 	}
 
@@ -53,10 +58,11 @@ public class EcoreServerRuntimeModule extends ServerModule {
 	public Class<? extends OperationConfiguration> bindOperationConfiguration() {
 		return EcoreOperationConfiguration.class;
 	}
-	
+
 	@Override
-	public Class<? extends OperationHandlerProvider> bindOperationHandlerProvider() {
+	
+	protected Class<? extends OperationHandlerProvider> bindOperatioHandlerProvider() {
 		return EcoreOperationHandlerProvider.class;
 	}
-	
+
 }
